pass printer via opts
